feat(security): Extract and enhance PasswordService into new Security module
