+++ apps/backend/users-service/src/auth/domain/interfaces/index.ts
@@ -1,3 +1,2 @@
-export * from './password.service.interface';
 export * from './refresh-token.repository.interface';
 export * from './token.service.interface';
