+++ apps/backend/users-service/src/auth/infrastructure/services/index.ts
@@ -1,2 +1,1 @@
-export * from './bcrypt-password.service';
 export * from './jwt-token.service';
