+++ apps/backend/users-service/src/security/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './password.service.interface';
