+++ apps/backend/users-service/src/security/interfaces/password.service.interface.ts
@@ -1,4 +1,5 @@
 export abstract class PasswordService {
   abstract hash(plain: string): Promise<string>;
   abstract compare(plain: string, hashed: string): Promise<boolean>;
+  abstract generateTemporary(length?: number): string;
 }
