+++ apps/backend/users-service/src/security/security.module.ts
@@ -1,0 +1,14 @@
+import { Module } from '@nestjs/common';
+import { PasswordService } from './interfaces';
+import { BcryptPasswordService } from './services';
+
+@Module({
+  providers: [
+    {
+      provide: PasswordService,
+      useClass: BcryptPasswordService,
+    },
+  ],
+  exports: [PasswordService],
+})
+export class SecurityModule {}
