+++ apps/backend/users-service/src/security/services/bcrypt-password.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import * as bcrypt from 'bcrypt';
-import { HASH_ROUNDS } from '../../../constants';
-import { PasswordService } from '../../domain';
+import * as crypto from 'crypto';
+import { HASH_ROUNDS } from '../constants';
+import { PasswordService } from '../interfaces';
 
 @Injectable()
 export class BcryptPasswordService implements PasswordService {
@@ -17,5 +18,15 @@ export class BcryptPasswordService implements PasswordService {
    */
   async compare(plain: string, hashed: string): Promise<boolean> {
     return bcrypt.compare(plain, hashed);
+  }
+
+  /**
+   * Generates a temporary password of specified length.
+   */
+  generateTemporary(length = 12): string {
+    return crypto
+      .randomBytes(Math.ceil(length / 2))
+      .toString('hex')
+      .slice(0, length);
   }
 }
