+++ apps/backend/users-service/src/security/services/index.ts
@@ -1,0 +1,1 @@
+export * from './bcrypt-password.service';
